Cache the rect sizing of nodes when we calculate it
When continually selecting text, the document will need to get the element
under the mouse cursor.  This results in a lot of layout calculatations
which slows down IE to a crawl.  Instead, cache those values so that if
we measure the same node multiple times without changing the document, we
can reuse the old values.

Fixes Issue #1